removed seeker collisions
