+++ Assets/Resources/Robber.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!114 &4617745966275864761
+--- !u!114 &1596534037847780934
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -9,7 +9,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 6809595694847979762}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f4c7357da2848435d8d6c2a7cc175496, type: 3}
+  m_Script: {fileID: 11500000, guid: 5fb2bf8859422d849b551bbfd8bf3d88, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1001 &6050403968897764889
