+++ Assets/Scripts/Interactables/Capturable.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Capturable : Interactable {
+  public void Capture() {
+    GameManager.instance.OnRobberCapture(gameObject);
+  }
+}
