+++ Assets/Scripts/Interactables/Capturable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5fb2bf8859422d849b551bbfd8bf3d88
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
