+++ Assets/Scripts/Interactables/Interactable.cs
@@ -2,9 +2,12 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-interface Interactable {
-  void PickUp(Transform pickUpDestination);
-  void PutDown();
-  void GlowOn();
-  void GlowOff();
+public abstract class Interactable : MonoBehaviour {
+  public void GlowOn() {
+    GetComponent<Outline>().enabled = true;
+  }
+
+  public void GlowOff() {
+    GetComponent<Outline>().enabled = false;
+  }
 }
