+++ Assets/Scripts/Interactables/PickUpable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f4c7357da2848435d8d6c2a7cc175496
+guid: f6ac5086691dd8236ad8243bd14d69b6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
